Added foreign key reference to teams table from countries table
The country attribute in teams table is now a foreign key which
references the abbreviation attribute in countries table. It is
restricted for delete operations but cascaded for update operations.

Also reset buttons are added which drops the table, creates it again and
enters some default values. The reset button for teams page won't work
if all the necessary countries don't exist in countries table. That
decision is optional, can be changed down the line.

Also modified the html and css files where necessary.