Added the final table and it's functions
Last table is a standings table for teams. It has 3 columns: position,
name and points. Position is the primary key. Name references the team
names in the teams table and names in the first table (teams) now have
the "UNIQUE" property. Similarly the names in the third table(standings)
also have the "UNIQUE" property. It cascades on update and delete. Table
is sorted by the position. 

It has add, update, delete and search functions. Update and delete are
based on position, users can search with name or points. 

Necessary html and py files are added.